Extract common functions and clean up includes
