+++ src/Parsers/Kusto/KustoFunctions/KQLAggregationFunctions.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include <Parsers/IParserBase.h>
-#include <Parsers/Kusto/KustoFunctions/IParserKQLFunction.h>
+#include "IParserKQLFunction.h"
+
 namespace DB
 {
 class ArgMax : public IParserKQLFunction
