+++ src/Parsers/Kusto/KustoFunctions/KQLFunctionFactory.cpp
@@ -1,20 +1,15 @@
-#include <Parsers/IParserBase.h>
-#include <Parsers/ParserSetQuery.h>
-#include <Parsers/ASTExpressionList.h>
-#include <Parsers/ASTSelectWithUnionQuery.h>
-#include <Parsers/Kusto/ParserKQLQuery.h>
-#include <Parsers/Kusto/ParserKQLStatement.h>
-#include <Parsers/Kusto/KustoFunctions/IParserKQLFunction.h>
-#include <Parsers/Kusto/KustoFunctions/KQLDateTimeFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLStringFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLDynamicFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLCastingFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLAggregationFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLTimeSeriesFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLIPFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLBinaryFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLGeneralFunctions.h>
-#include <Parsers/Kusto/KustoFunctions/KQLFunctionFactory.h>
+#include "KQLFunctionFactory.h"
+
+#include "IParserKQLFunction.h"
+#include "KQLDateTimeFunctions.h"
+#include "KQLStringFunctions.h"
+#include "KQLDynamicFunctions.h"
+#include "KQLCastingFunctions.h"
+#include "KQLAggregationFunctions.h"
+#include "KQLTimeSeriesFunctions.h"
+#include "KQLIPFunctions.h"
+#include "KQLBinaryFunctions.h"
+#include "KQLGeneralFunctions.h"
 
 namespace DB
 {
